+++ src/c_cmds.c
@@ -4398,10 +4398,6 @@ static void mapstats_cmd_func2(char *cmd, char *parms)
     C_TabbedOutput(tabs, INDENT "Damaging\t%s", temp);
     free(temp);
 
-    temp = commify(nummovable);
-    C_TabbedOutput(tabs, INDENT "Movable\t%s", temp);
-    free(temp);
-
     if (blockmaprebuilt)
         C_TabbedOutput(tabs, "Blockmap\tRebuilt");