+++ src/p_ceilng.c
@@ -363,8 +363,6 @@ void P_AddActiveCeiling(ceiling_t *ceiling)
 
     list->prev = &activeceilings;
     activeceilings = list;
-
-    nummovable++;
 }
 
 //