+++ src/p_plats.c
@@ -360,8 +360,6 @@ void P_AddActivePlat(plat_t *plat)
 
         list->prev = &activeplats;
         activeplats = list;
-
-        nummovable++;
     }
 }